Sample update, K8S registry extension (#430)
