+++ dubbo-samples-kubernetes/dubbo-samples-apiserver-consumer/src/main/java/org/apache/dubbo/samples/action/GreetingServiceConsumer.java
@@ -20,14 +20,14 @@
 package org.apache.dubbo.samples.action;
 
 import org.apache.dubbo.config.annotation.DubboReference;
-import org.apache.dubbo.config.annotation.Reference;
 import org.apache.dubbo.samples.api.GreetingService;
+
 import org.springframework.stereotype.Component;
 
 @Component("annotatedConsumer")
 public class GreetingServiceConsumer {
 
-    @DubboReference(version = "1.0.0", services = {"kubernetes-apiserver-demo-provider"})
+    @DubboReference(version = "1.0.0", providedBy = "dubbo-samples-apiserver-provider")
     private GreetingService greetingService;
 
     public String doSayHello(String name) {
