+++ dubbo-samples-kubernetes/dubbo-samples-apiserver-consumer/src/main/resources/k8s/Service.yml
@@ -1,0 +1,13 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: dubbo-samples-apiserver-consumer
+  namespace: dubbo-demo
+spec:
+  clusterIP: None
+  selector:
+    app: dubbo-samples-apiserver-consumer
+  ports:
+    - protocol: TCP
+      port: 20880
+      targetPort: 20880
