+++ dubbo-samples-kubernetes/dubbo-samples-apiserver-consumer/src/main/resources/log4j.properties
@@ -16,7 +16,6 @@
 #   limitations under the License.
 #  
 #
-
 log4j.rootLogger=info, stdout
 log4j.appender.stdout=org.apache.log4j.ConsoleAppender
 log4j.appender.stdout.Target=System.out
