+++ dubbo-samples-kubernetes/dubbo-samples-apiserver-consumer/src/main/resources/spring/dubbo-consumer.properties
@@ -16,10 +16,9 @@
 #   limitations under the License.
 #
 #
-
-dubbo.application.name=kubernetes-apiserver-demo-consumer
+dubbo.application.name=dubbo-samples-apiserver-consumer
 dubbo.application.metadataServicePort=20885
-dubbo.registry.address=kubernetes://${your kubernetes api server ip here}:${your kubernetes api server port here}?registry-type=service&duplicate=false&namespace=dubbo-demo&trustCerts=true&oauthToken=${your ServiceAccount token here}
+dubbo.registry.address=kubernetes://DEFAULT_MASTER_HOST?registry-type=service&duplicate=false&namespace=dubbo-demo&trustCerts=true&subscribed-services=dubbo-samples-apiserver-provider
 dubbo.consumer.timeout=3000
 dubbo.application.qosEnable=true
 dubbo.application.qosAcceptForeignIp=true
