+++ dubbo-samples-kubernetes/dubbo-samples-apiserver-provider/src/main/resources/spring/dubbo-provider.properties
@@ -16,10 +16,9 @@
 #   limitations under the License.
 #
 #
-
-dubbo.application.name=kubernetes-apiserver-demo-provider
+dubbo.application.name=dubbo-samples-apiserver-provider
 dubbo.application.metadataServicePort=20885
-dubbo.registry.address=kubernetes://${your kubernetes api server ip here}:${your kubernetes api server port here}?registry-type=service&duplicate=false&namespace=dubbo-demo&trustCerts=true&oauthToken=${your ServiceAccount token here}
+dubbo.registry.address=kubernetes://DEFAULT_MASTER_HOST?registry-type=service&duplicate=false&namespace=dubbo-demo&trustCerts=true
 dubbo.protocol.name=dubbo
 dubbo.protocol.port=20880
 dubbo.application.qosEnable=true
