+++ dubbo-samples-kubernetes/dubbo-samples-kubernetes-dns/src/main/java/org/apache/dubbo/samples/ConsumerBootstrap.java
@@ -21,6 +21,7 @@ package org.apache.dubbo.samples;
 
 import org.apache.dubbo.config.spring.context.annotation.EnableDubbo;
 import org.apache.dubbo.samples.action.GreetingServiceConsumer;
+
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
