+++ dubbo-samples-kubernetes/dubbo-samples-kubernetes-dns/src/main/java/org/apache/dubbo/samples/ProviderBootstrap.java
@@ -20,6 +20,7 @@
 package org.apache.dubbo.samples;
 
 import org.apache.dubbo.config.spring.context.annotation.EnableDubbo;
+
 import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.PropertySource;
