+++ dubbo-samples-kubernetes/dubbo-samples-kubernetes-dns/src/main/resources/spring/dubbo-consumer.properties
@@ -16,7 +16,6 @@
 #   limitations under the License.
 #
 #
-
 dubbo.application.name=kubernetes-dns-demo-consumer
 dubbo.application.metadataServicePort=20885
 dubbo.registry.address=dns://${your kube dns ip here}:53?registry-type=service&duplicate=false&addressSuffix=.dubbo-demo.svc.cluster.local.
