+++ dubbo-samples-kubernetes/pom.xml
@@ -24,8 +24,9 @@
     <groupId>org.apache.dubbo</groupId>
     <version>1.0-SNAPSHOT</version>
     <modules>
-        <module>dubbo-samples-kubernetes-apiserver</module>
-        <module>dubbo-samples-kubernetes-dns</module>
+        <module>dubbo-samples-apiserver-provider</module>
+        <module>dubbo-samples-apiserver-consumer</module>
+        <!--        <module>dubbo-samples-kubernetes-dns</module>-->
     </modules>
 
     <modelVersion>4.0.0</modelVersion>
